requirements.txt
setuptools<81
django==5.1
djangorestframework==3.15
psycopg[binary]>=3.1
dj-database-url==2.2
python-decouple==3.8
djangorestframework-simplejwt==5.3
django-cors-headers==4.4
whitenoise==6.7
celery==5.4
redis==5.0
channels==4.1
channels-redis==4.2
daphne==4.1
anthropic
twilio==9.3
supabase
Pillow
python-magic
django-celery-beat==2.7
requests
gunicorn==22.0.0
pytz

groq
